Use fix prefix for external USB network device names on compact
Drivers like the realtech r8152 name their interfaces "ethX" (with X being
the number of the interface). This may collide with the internal USB and
SPI ethernet interfaces "eth0" and "eth1" on the compact and then cause a
rename of the concerning internal interface. To avoid this add a udev rule
to rename all network interfaces that are connected to the (external) USB
with the name prefix "usbnet".

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>